Separate Debian 11 into its own scenario
We need to not test Debian 11 for idempotence.  This is because
aws/efs-utils currently breaks idempotence for that platform.  This is
because it forces the installation of Python2.  When the idempotence
check runs it therefore tries to install python-apt, which does not
exist for Debian 11.

I have created aws/efs-utils#57 to encourage them to support python3
in their packaging.  (The code itself already supports python3.)