Add a force_python3 inventory group for Kali and Debian 11
auto_legacy is still the default behavior until Ansible 2.12 is
released, but Molecule overrides this and forces the use of auto.

In auto mode Ansible will prefer the Python interpreter specified in
an internal map.  If there is no entry in the internal map then
Ansible will search an ordered list of potential Python locations and
use the first Python executable from that list that it finds.  At the
top of that list is /usr/bin/python.

In most cases the internal map specifies /usr/bin/python3, but neither
Debian 11 nor Kali Linux are listed in the internal map.  Since the
installation of amazon-efs-utils forces the installation of Python
2 (located at /usr/bin/python), this means that Ansible ends up using
/usr/bin/python3 for the converge playbook but /usr/bin/python for the
idempotence playbook.  This causes those two platforms to fail
idempotence.

To get around this we force the use of /usr/bin/python3 for these two
platforms.