Use 1.x version of amazon-efs-utils on Ubuntu Jammy
The Molecule tests pass when we use version 2.x on that platform, but
when building an actual AMI the installation of cargo pulls in the
libssh2-1 package, which seems to break the running ssh server.  See
this build, for example:
https://github.com/cisagov/ubuntu-server-packer/actions/runs/9226352690/job/25386015270?pr=24

On Ubuntu Noble the installation of cargo pulls in libssh2-1t64, which
hopefully does not break the running ssh server.